requirements.txt
streamlit
PyPDF2
langchain_openai
langchain_community
langchain
langchain_core
python-dotenv
